implement websocket endpoint as a seperate application
need to place it in a different application to make it run
Why? After adding it to the main application, I couldn't reach no normal url e.g. http://localhost:8080/replay/view/{id}
Now you'll need to deploy that app to a container and
let it run in "replays" context. This will probably change